🔖 1.8.1: ✨ Add `pagination-by-month` to `dashboard GUI`
