+++ src/lib/util.ts
@@ -49,3 +49,59 @@ export class AutoId {
     return autoId
   }
 }
+
+export function printMonthYear (d: Date): string {
+  const options = {
+    year: 'numeric',
+    month: 'long'
+  }
+  return new Intl.DateTimeFormat('en-US', options as any).format(d)
+}
+
+export function getNextMonth (d: Date): Date {
+  const nd = new Date(d.getFullYear(), d.getMonth() + 1, d.getDate()) // Step forward one month
+  // console.log('>> getNextMonth:', d.toDateString(), nd.toDateString())
+  return nd
+}
+
+export function getPrevMonth (d: Date): Date {
+  const nd = new Date(d.getFullYear(), d.getMonth() - 1, 1) // Step backward one month; start on first day
+  // console.log('>> getPrevMonth:', d.toDateString(), nd.toDateString())
+  return nd
+}
+
+export function getPeriod (d: Date): string {
+  return `${d.getFullYear()}-` + `${d.getMonth() + 1}`.padStart(2, '0')
+}
+
+export function getMonthLastDay (d: Date): Date {
+  const nd = new Date(d.getFullYear(), d.getMonth() + 1, 0)
+  return nd
+}
+
+export type TDay = {
+  dayNo: number
+  dateStr: string
+}
+
+export function getDateRange (endDate: Date): TDay[] {
+  const today = new Date()
+  let dateCursor: Date
+  if (today.getMonth() === endDate.getMonth()) {
+    dateCursor = new Date(endDate.getFullYear(), endDate.getMonth(), today.getDate()) // Start on today
+  } else {
+    dateCursor = new Date(endDate.getFullYear(), endDate.getMonth() + 1, 0) // Start on the last day of the month
+  }
+
+  const dateRange = [] as TDay[]
+  const startDate = new Date(endDate.getFullYear(), endDate.getMonth(), 1) // Stop on the first day of the month at midnight ET
+  // Eg. Tue Jun 01 2021 00:00:00 GMT-0400 (Eastern Daylight Time)
+  while (dateCursor.getTime() >= startDate.getTime()) {
+    dateRange.push({
+      dayNo: dateCursor.getDay(),
+      dateStr: dateCursor.toLocaleDateString()
+    })
+    dateCursor.setTime(dateCursor.getTime() - 86400 * 1000) // Step one day backwards until we get to the start date
+  }
+  return dateRange
+}
